fix golangci lint problom

+++ pkg/encoding/encoding.go
@@ -3,7 +3,7 @@ package encoding
 // ZigZagEncode converts a int64 to a uint64 by zig zagging negative and positive values
 // across even and odd numbers.  Eg. [0,-1,1,-2] becomes [0, 1, 2, 3].
 func ZigZagEncode(x int64) uint64 {
-	return uint64(uint64(x<<1) ^ uint64(int64(x) >> 63))
+	return uint64(x<<1) ^ uint64(x>>63)
 }
 
 // ZigZagDecode converts a previously zigzag encoded uint64 back to a int64.
